parse prefatory content as Presentation XML, have it use collection bibdata, merged with any bibdata added in semantic XML processing; use preface of prefatory content if sections are empty: https://github.com/metanorma/metanorma/issues/512
