+++ spec/collection/collection_complex_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require_relative "../spec_helper"
 require "stringio"
 
 def capture_stdout
