+++ spec/fixtures/collection/collection1.xml
@@ -288,7 +288,7 @@
   <format>html</format>
   <coverpage>cover.html</coverpage>
           <prefatory-content>
-             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="2">
+             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="3">
                 <title id="_">Clause</title>
                 <fmt-title depth="1" id="_">
                    <semx element="title" source="_">Clause</semx>
@@ -297,7 +297,7 @@
              </clause>
           </prefatory-content>
           <final-content>
-             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="2">
+             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="3">
                 <title id="_">Exordium</title>
                 <fmt-title depth="1" id="_">
                    <semx element="title" source="_">Exordium</semx>
