+++ spec/fixtures/collection/collection_docinline.xml
@@ -281,7 +281,7 @@
   <format>html</format>
   <coverpage>cover.html</coverpage>
           <prefatory-content>
-             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="2">
+             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="3">
                 <title id="_">Clause</title>
                 <fmt-title depth="1" id="_">
                    <semx element="title" source="_">Clause</semx>
@@ -533,7 +533,7 @@
     <attachment>data:image/svg+xml</attachment>
   </doc-container>
           <final-content>
-             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="2">
+             <clause id="_" inline-header="false" obligation="normative" unnumbered="true" displayorder="3">
                 <title id="_">Exordium</title>
                 <fmt-title depth="1" id="_">
                    <semx element="title" source="_">Exordium</semx>
