+++ spec/fixtures/collection/cover-iho.html
@@ -1,0 +1,2 @@
+<p>Intro Document {{ doctitle }} {{ docnumeric }}</p>
+{{ prefatory-content }}
