+++ spec/fixtures/collection/iho.xml
@@ -1,0 +1,116 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metanorma xmlns="https://www.metanorma.org/ns/standoc" type="semantic" version="1.6.0" flavor="iho">
+<bibdata type="standard">
+<title language="en" format="text/plain" type="main">Test conditions for milling machines with table of variable height, with horizontal or vertical spindle</title>
+<title language="en" format="text/plain" type="title-main">Test conditions for milling machines with table of variable height, with horizontal or vertical spindle</title>
+<title language="fr" format="text/plain" type="main">Conditions d’essais des machines à fraiser à table de hauteur variable, à broche horizontale ou verticale</title>
+<title language="fr" format="text/plain" type="title-main">Conditions d’essais des machines à fraiser à table de hauteur variable, à broche horizontale ou verticale</title>
+<docidentifier primary="true" type="ISO">ISO 1701<sup>x</sup>:1974</docidentifier><docidentifier type="iso-with-lang">ISO 1701:1974(E)</docidentifier><docidentifier type="iso-reference">ISO 1701:1974(E)</docidentifier>
+
+
+<docnumber>1701</docnumber>
+<contributor>
+<role type="author"/>
+<organization>
+<name>International Organization for Standardization</name>
+<abbreviation>ISO</abbreviation>
+</organization>
+</contributor>
+<contributor>
+<role type="publisher"/>
+<organization>
+<name>International Organization for Standardization</name>
+<abbreviation>ISO</abbreviation>
+</organization>
+</contributor>
+<edition>1</edition>
+<version>
+<revision-date>2016-05-01</revision-date>
+</version>
+<language>en</language>
+<script>Latn</script>
+<status>
+<stage abbreviation="IS">60</stage>
+<substage>60</substage>
+</status>
+<copyright>
+<from>1974</from>
+<owner>
+<organization>
+<name>International Organization for Standardization</name>
+<abbreviation>ISO</abbreviation>
+</organization>
+</owner>
+</copyright>
+<ext>
+<doctype>international-standard</doctype>
+<ics>
+<code>67.060</code>
+</ics>
+<structuredidentifier>
+<project-number>ISO 1701</project-number>
+</structuredidentifier>
+<stagename>International standard</stagename>
+</ext>
+</bibdata>
+<boilerplate>
+  <copyright-statement>
+    <clause>
+      <title>COPYRIGHT PROTECTED DOCUMENT</title>
+    <p id="boilerplate-year">
+    © ISO 1974
+  </p>
+
+  <p id="boilerplate-message">
+    All rights
+reserved. Unless otherwise specified, no part of this publication may be
+reproduced or utilized otherwise in any form or by any means, electronic or
+mechanical, including photocopying, or posting on the internet or an intranet,
+without prior written permission. Permission can be requested from either ISO
+at the address below or ISO’s member body in the country of the requester.
+  </p>
+
+  <p id="boilerplate-address" align="left">
+    ISO copyright office<br/>
+    Ch. de Blandonnet 8 • CP 401<br/>
+    CH-1214 Vernier, Geneva, Switzerland<br/>
+    Tel.  + 41 22 749 01 11<br/>
+    Fax  + 41 22 749 09 47<br/>
+    Email: copyright@iso.org<br/>
+    Website: www.iso.org
+  </p>
+    <p id="boilerplate-place">
+    Published in Switzerland
+  </p>
+</clause>
+  </copyright-statement>
+
+  
+</boilerplate>
+<sections><clause id="_scope" type="scope" inline-header="false" obligation="normative"><title>Scope</title><p id="_0f24bcbe-6ca7-400e-9fb3-34027510f8f2">This document specifies minimum requirements and test methods for rice (<em>Oryza sativa L.</em>).</p>
+<p id="_7b4ddc83-6eab-4476-99ab-dee2fa285d14">It is applicable to husked rice, husked parboiled rice, milled rice and milled parboiled rice, suitable for human consumption, directly or after reconditioning.</p>
+<p id="_2d5119f1-3fed-4faa-a457-65bb0c0449d2">It is not applicable to cooked rice products.</p>
+<p id="_3ca0ca2e-49cb-45c2-9bd5-e79bdaae0f7a">This document is updated in <eref type="inline" bibitemid="RiceAmd" citeas="ISO 17301-1:2016/Amd.1:2017"/>. It derives from <eref type="inline" bibitemid="StarTrek" citeas="ISO 1701:1974"/>. See e.g. <eref type="inline" bibitemid="StarTrek" citeas="ISO 1701:1974"><localityStack><locality type="clause"><referenceFrom>1</referenceFrom></locality></localityStack></eref>.</p>
+<p id="_d726ee55-0554-49a3-851e-dc8e16372c3f">This document is unrelated to <eref type="inline" bibitemid="Dummy" citeas="ISO 17302:2016"><localityStack><locality type="anchor"><referenceFrom>what</referenceFrom></locality></localityStack></eref>.</p></clause>
+</sections><bibliography><references id="_bibliography" normative="false" obligation="informative">
+<title>Bibliography</title><bibitem id="StarTrek">
+<formattedref format="application/x-isodoc+xml"/>
+<docidentifier type="ISO">ISO 1701:1974</docidentifier>
+<docidentifier type="repository">current-metanorma-collection/rice</docidentifier>
+<docnumber>1701:1974</docnumber>
+</bibitem><bibitem id="RiceAmd">
+<formattedref format="application/x-isodoc+xml"/>
+<docidentifier type="ISO">ISO 17301-1:2016/Amd.1:2017</docidentifier>
+<docidentifier type="repository">current-metanorma-collection/ISO 17301-1:2016/Amd.1:2017</docidentifier>
+<docnumber>17301-1:2016/Amd.1:2017</docnumber>
+</bibitem><bibitem id="Dummy">
+<formattedref format="application/x-isodoc+xml"/>
+<docidentifier type="ISO">ISO 17302:2016</docidentifier>
+<docidentifier type="repository">current-metanorma-collection/ISO 17302<sup>x</sup></docidentifier>
+<docnumber>17302</docnumber>
+</bibitem>
+
+
+
+</references></bibliography>
+</metanorma>
